Change data type id to BigInt

+++ src/main/java/com/example/happy_fitness/common/RoleEnum.java
@@ -2,19 +2,21 @@ package com.example.happy_fitness.common;
 
 import lombok.Getter;
 
+import java.math.BigInteger;
+
 @Getter
 public enum RoleEnum {
-    ROLE_ADMIN(1F, "Admin"),
-    ROLE_MANAGER(2F, "Manager"),
-    ROLE_CUSTOMER(3F, "Customer"),
-    ROLE_RECEPTIONIST(4F, "Receptionist"),
-    ROLE_PERSONAL_TRAINER(5F, "Personal Trainer");
+    ROLE_ADMIN(BigInteger.valueOf(1), "Admin"),
+    ROLE_MANAGER(BigInteger.valueOf(2), "Manager"),
+    ROLE_CUSTOMER(BigInteger.valueOf(3), "Customer"),
+    ROLE_RECEPTIONIST(BigInteger.valueOf(4), "Receptionist"),
+    ROLE_PERSONAL_TRAINER(BigInteger.valueOf(5), "Personal Trainer");
 
-    private final Float id;
+    private final BigInteger id;
 
     private final String value;
 
-    RoleEnum(Float id, String value) {
+    RoleEnum(BigInteger id, String value) {
         this.id = id;
         this.value = value;
     }
