+++ src/main/java/com/example/happy_fitness/dto/CartDto.java
@@ -2,10 +2,12 @@ package com.example.happy_fitness.dto;
 
 import lombok.Data;
 
+import java.math.BigInteger;
+
 @Data
 public class CartDto {
-    private Float id;
-    private Float facilityProductId;
+    private BigInteger id;
+    private BigInteger facilityProductId;
     private String code;
     private String name;
     private String image;
