+++ src/main/java/com/example/happy_fitness/repository/CartRepository.java
@@ -4,7 +4,9 @@ import com.example.happy_fitness.entity.Cart;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigInteger;
+
 @Repository
-public interface CartRepository extends JpaRepository<Cart, Float> {
-    Cart findByCreatedByAndFacilityProduct_Id(String username, Float facilityProductId);
+public interface CartRepository extends JpaRepository<Cart, BigInteger> {
+    Cart findByCreatedByAndFacilityProduct_Id(String username, BigInteger facilityProductId);
 }
