+++ src/main/java/com/example/happy_fitness/repository/CategoryRepository.java
@@ -4,6 +4,8 @@ import com.example.happy_fitness.entity.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigInteger;
+
 @Repository
-public interface CategoryRepository extends JpaRepository<Category, Float> {
+public interface CategoryRepository extends JpaRepository<Category, BigInteger> {
 }
