+++ src/main/java/com/example/happy_fitness/repository/CustomerTicketRepository.java
@@ -4,13 +4,14 @@ import com.example.happy_fitness.entity.CustomerTicket;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigInteger;
 import java.util.List;
 
 @Repository
-public interface CustomerTicketRepository extends JpaRepository<CustomerTicket, Float> {
+public interface CustomerTicketRepository extends JpaRepository<CustomerTicket, BigInteger> {
     List<CustomerTicket> findAllByStatusIsTrue();
 
-    List<CustomerTicket> findAllByCustomer_IdAndStatusIsTrue(Float customerId);
+    List<CustomerTicket> findAllByCustomer_IdAndStatusIsTrue(BigInteger customerId);
 
     List<CustomerTicket> findAllByCustomer_UsernameOrderByUpdatedDateDesc(String username);
 }
