+++ src/main/java/com/example/happy_fitness/repository/FacilityRepository.java
@@ -3,5 +3,7 @@ package com.example.happy_fitness.repository;
 import com.example.happy_fitness.entity.Facility;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FacilityRepository extends JpaRepository<Facility, Float> {
+import java.math.BigInteger;
+
+public interface FacilityRepository extends JpaRepository<Facility, BigInteger> {
 }
