+++ src/main/java/com/example/happy_fitness/service/OrderService.java
@@ -5,12 +5,13 @@ import com.example.happy_fitness.dto.OrderDto;
 import com.example.happy_fitness.entity.Order;
 import org.springframework.security.core.userdetails.UserDetails;
 
+import java.math.BigInteger;
 import java.util.List;
 
 public interface OrderService extends BaseService<Order> {
-    String order(List<Float> carts, String voucherCode);
+    String order(List<BigInteger> carts, String voucherCode);
 
     List<OrderDto> findOrders(UserDetails userDetails);
 
-    List<OrderDetailDto> findOrderDetail(Float id, UserDetails userDetails);
+    List<OrderDetailDto> findOrderDetail(BigInteger id, UserDetails userDetails);
 }
