+++ src/main/java/com/example/happy_fitness/service/ProductService.java
@@ -6,16 +6,17 @@ import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
+import java.math.BigInteger;
 import java.util.List;
 
 public interface ProductService extends BaseService<Product> {
-    List<ProductDto> findProducts(Float facilityId, String status, Float categoryId, Float supplierId, Float price, Float maxPrice);
+    List<ProductDto> findProducts(BigInteger facilityId, String status, BigInteger categoryId, BigInteger supplierId, Float price, Float maxPrice);
 
-    ProductDto findProductDetail(Float facilityId, String code);
+    ProductDto findProductDetail(BigInteger facilityId, String code);
 
     String createCustom(UserDetails userDetails, Product product, MultipartFile image) throws IOException;
 
-    String updateCustom(Product product, Float id, MultipartFile image) throws IOException;
+    String updateCustom(Product product, BigInteger id, MultipartFile image) throws IOException;
 
-    void active(Float id);
+    void active(BigInteger id);
 }
